z buffer finally works fine

+++ src/animation/lightzbufferrenderer.h
@@ -24,7 +24,7 @@ public:
 
 protected:
     void fillTriangle(Triangle &triangle);
-    void getLeftRightBounds(std::vector<int> &lleft, std::vector<int> &lright,
+    void getLeftRightBounds(std::vector<float> &lleft, std::vector<float> &lright,
                             std::vector<Vec3> &nleft, std::vector<Vec3> &nright, std::vector<float> &zleft,
                             std::vector<float> &zright,
                             Triangle &triangle,
@@ -33,8 +33,8 @@ protected:
     float calculateIntensity(const Vec3 &n, const Vec3 &orig);
 
     void putPixel(int x, int y, const Color &color);
-    std::vector<int> getBrezenhemX(const Vec3 &p1, const Vec3 &p2);
-    std::vector<Vec3> getNormals(const std::vector<int> &l, const Vec3 &n1,
+    std::vector<float> getBrezenhemX(const Vec3 &p1, const Vec3 &p2);
+    std::vector<Vec3> getNormals(const std::vector<float> &l, const Vec3 &n1,
                                  const Vec3 &n2);
     std::vector<float> getZLine(const Vec3 &p1, const Vec3 &p2, int n);
